refactor: have types locally
