--- ts/logic/inputs/index.ts
@@ -1,7 +1,0 @@
-export * from "./factory";
-export * from "./InputBase";
-export * from "./InputFirstRow";
-export * from "./InputGap";
-export * from "./InputMediaQuery";
-export * from "./InputNumberOfHexagons";
-export * from "./InputRoot";
